WIP: Evacuator proof of concept (addendum to fix for issue #1)
Fixes: #1 (Regression in 68e1176b)

This eliminates residual excessive stf/ttf counts in trace output
remaining after previous commit was applied. These occurred rarely but
always when the inside and outside copyspaces were too full to receive
copy but not full enough to allow refreshment, because primitive objects
were not being admitted for outside copy which would force outside
copyspace to be refreshed.

This commit also establishes an efficient way to direct object copy
while scanning in topmost frame. This method was present in the previous
commit but is refined here. For all I know this will change again if and
when I reverse the order in which application thread stack roots are
evacuated (currently top-down, which is like parsing an XML document
while reading it backwards).

See comments in issue #1 for benchmarking results with this commit.